+++ app/src/main/java/cl/agnov/ameli/ui/navigation/AmeliNavHost.kt
@@ -1,14 +1,23 @@
 package cl.agnov.ameli.ui.navigation
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import cl.agnov.ameli.AmeliApplication
+import cl.agnov.ameli.sip.model.CallConnectionState
+import cl.agnov.ameli.sip.model.CallDirection
 import cl.agnov.ameli.ui.screens.ActiveCallScreen
 import cl.agnov.ameli.ui.screens.DialerScreen
 import cl.agnov.ameli.ui.screens.HomeScreen
+import cl.agnov.ameli.ui.screens.IncomingCallScreen
 import cl.agnov.ameli.ui.screens.SettingsScreen
 
 object AmeliDestinations {
@@ -16,6 +25,7 @@ object AmeliDestinations {
     const val SETTINGS = "settings"
     const val DIALER = "dialer"
     const val ACTIVE_CALL = "active_call"
+    const val INCOMING_CALL = "incoming_call"
 }
 
 @Composable
@@ -23,6 +33,19 @@ fun AmeliNavHost(
     modifier: Modifier = Modifier,
     navController: NavHostController = rememberNavController(),
 ) {
+    val application = LocalContext.current.applicationContext as AmeliApplication
+    val callState by application.container.callManager.callState.collectAsState()
+    val currentRoute by navController.currentBackStackEntryAsState()
+
+    LaunchedEffect(callState?.connectionState, currentRoute) {
+        val isIncomingRinging = callState?.direction == CallDirection.INCOMING &&
+            callState?.connectionState == CallConnectionState.INCOMING_RINGING
+        val alreadyShowingIncomingCall = currentRoute?.destination?.route == AmeliDestinations.INCOMING_CALL
+        if (isIncomingRinging && !alreadyShowingIncomingCall) {
+            navController.navigate(AmeliDestinations.INCOMING_CALL)
+        }
+    }
+
     NavHost(
         navController = navController,
         startDestination = AmeliDestinations.HOME,
@@ -47,6 +70,18 @@ fun AmeliNavHost(
         composable(AmeliDestinations.ACTIVE_CALL) {
             ActiveCallScreen(
                 onCallEnded = {
+                    navController.popBackStack(AmeliDestinations.HOME, inclusive = false)
+                },
+            )
+        }
+        composable(AmeliDestinations.INCOMING_CALL) {
+            IncomingCallScreen(
+                onAnswered = {
+                    navController.navigate(AmeliDestinations.ACTIVE_CALL) {
+                        popUpTo(AmeliDestinations.INCOMING_CALL) { inclusive = true }
+                    }
+                },
+                onDismissed = {
                     navController.popBackStack(AmeliDestinations.HOME, inclusive = false)
                 },
             )